pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hcai-discover-utils"
authors = [
    {name = "Dominik Schiller", email = "dominik.schiller@uni-a.de"}
]
description = "This repository contains utility functions and interfaces that can be used to interact with the NOVA annotation tool."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["NOVA", "utility", "annotation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only"
]
dependencies = [
    "ffmpegio >= 0.8.3",
    "pims",
    "soundfile",
    "imageio",
    "imageio-ffmpeg",
    "numba >= 0.57.1",
    "numpy >= 1.24",
    "pymongo == 3.12.2",
    "python-dotenv == 1.0.0",
    "requests",
    "pillow"
]
dynamic = ["version"]


[project.optional-dependencies]
decord = ["eva-decord >= 0.6.1"]
pymovie = ["pymovie"]
pyav = ["pyav"]



[tool.setuptools.packages.find]
#where = ["src"]  # list of folders that contain the packages (["."] by default)
#include = ["my_package*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["*tests*"]  # exclude packages matching these glob patterns (empty by default)

[tool.setuptools.dynamic]
version = {attr = "nova_utils.__version__"}

[project.urls]
Documentation = "https://hcmlab.github.io/discover-utils/docbuild/"
Repository = "https://github.com/hcmlab/discover-utils"

[project.scripts]
du-train = "discover_utils.scripts.train:cl_main"
du-process = "discover_utils.scripts.process:cl_main"


